// Order_Management_MS/src/service/order.service.ts
import { InProcessOrder, OrderLineItemType, OrderWithLineItems } from "../dto/orderRequest.dto";
import { CartRepositoryType } from "../repository/cart.repository";
import { OrderRepositoryType } from "../repository/order.repository";
import { MessageType } from "../types";
import { OrderStatus } from "../types/order.type";
import { SendCreateOrderMessage } from "./broker.service";

export const CreateOrder = async (
    userId: number,
    repo: OrderRepositoryType,
    cartRepo: CartRepositoryType
) => {
    // find cart by customer id
    const cart = await cartRepo.findCart(userId);
    if(!cart) {
        throw new Error("Cart not found");
    }

    // calculate total order amount
    let cartTotal = 0;
    let orderLineItems: OrderLineItemType[] = [];


    // create orderLineItems from cart items
    cart.lineItems.forEach((item)=> {
        cartTotal += item.qty * Number(item.price);
        orderLineItems.push({
            productId: item.productId,
            itemName: item.itemName,
            qty: item.qty,
            price: item.price
        } as OrderLineItemType)
    });
    
    // create order with line items

    const orderNumber =  Math.floor(Math.random() * 100000);
    const orderInput: OrderWithLineItems = {
        orderNumber: orderNumber,
        txnId: null,
        status: OrderStatus.PENDING,
        customerId: userId,
        amount: cartTotal.toString(),
        orderItems: orderLineItems
    }

    const order = await repo.createOrder(orderInput);
    await cartRepo.clearCartData(userId);
    console.log("Order created ", order)
    // fire a message to catalog service to update stock
    await SendCreateOrderMessage(orderInput);

    // return success message
    return { message: "Order created successfully", orderNumber: orderNumber};
};

export const UpdateOrder = async (
    orderId: number,
    status: OrderStatus,
    repo: OrderRepositoryType
) => {

    await repo.updateOrder(orderId, status);

    // Fire a message to catalog service to update stock
    if(status === OrderStatus.CANCELLED) {
        // await repo.publishOrderEvent(order, "ORDER_CANCELLED");
    }
    return { message : "Order updated successfully"}
};

export const GetOrder = (
    orderId: number,
    repo: OrderRepositoryType
) => {

    const order = repo.findOrder(orderId);
    if(!order) {
        throw new Error("Order not found")
    }

    return order;
};

export const GetOrders = async(
    userId: number,
    repo: OrderRepositoryType
) => {
    const orders = await repo.findOrdersByCustomerId(userId);
    if(!Array.isArray(orders)) {
        throw new Error("Orders not found")
    }
    return orders;
};

export const DeleteOrder = async(
    orderId: number,
    repo: OrderRepositoryType
) => {
    await repo.deleteOrder(orderId);
    return true;
};

export const HandleSubscription = (message: MessageType) => {

    // Here handle payment service messages for update order status

    console.log("Message received from consumer", message);
    // if message.event === OrderEvent.ORDER_UPDATED
    // call create order
}

export const CheckoutOrder = async( orderNumber: number, repo: OrderRepositoryType) => {
    const order = await repo.findOrderByNumber(orderNumber);
    if(!order){
        throw new Error("Order not found")
    }

    const checkoutOrder: InProcessOrder = {
        id: order.id,
        orderNumber: order.orderNumber,
        status: order.status,
        customerId: order.customerId,
        amount: Number(order.amount),
        createdAt: order.createdAt,
        updatedAt: order.updatedAt,
    }

    return checkoutOrder;
}